manager/requirements.txt
docker
pyyaml
websockets